Split code up into files
