+++ packages/graphile-utils/src/AddInflectorsPlugin.js
@@ -1,0 +1,7 @@
+export default function AddInflectorsPlugin(additionalInflectors) {
+  return builder => {
+    builder.hook("inflection", (inflection, build) => {
+      return build.extend(inflection, additionalInflectors);
+    });
+  };
+}
